Initial commit
In theory, there is no place for ingestion from a second source for a source-specific ingestion process.  But we are prototyping!

osc-ingest-DPDHL is derived from osc-ingest-shell, and works on the DPDHL xlsx spreadsheet in the data/raw directory.  Over time the goal is to unify functionality from DPDHL back into Shell without breaking Shell's processing of Shell's data.  When we have a single notebook able to handle both input spreadsheets in a unified fashion, we should start a new generic ingestion pipeline which can then be branched to ingest yet new similar formats, merging down to the base until we have a single script that can handle dozens if not hundreds of similarly-shaped files.

Note that at 10 minutes/NLP processing step, 150 files = 1000 minutes = approx 1 CPU day.  This script can probably handle 150 files in a minute or two, which is much more carbon-friendly.