Another WIP progress checkin
Biggest change is that we now remove notes by reading the XML.  We still need to do a better job of stashing references to those notes, but things like having a superscripted 7 on the year 2019 no longer comes back as 20197 (which is impossible to fix otherwise).

But lots of additional refactoring, unit handling, and slightly more to go.